backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
sqlalchemy==2.0.32
pymysql==1.1.1
alembic==1.13.2
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2
google-generativeai==0.7.2
razorpay==1.4.2
waitress==3.0.0
